+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.7.45
 
 * more efficient returning in `enumerate`
+* add `measure.has_exceeded`
+